Refactored from old calendars controller to 3 controllers: auth, calendars, events
